requirements.txt
python-dotenv~=1.0.0
pydub~=0.25.1
tiktoken==0.7.0
python-telegram-bot==21.9
requests~=2.32.3
tenacity==8.3.0
whois~=0.9.27
